add new satp3 policy
