+++ src/schedlib/policies/__init__.py
@@ -1,4 +1,5 @@
 from .flex import FlexPolicy
 from .sat import SATPolicy
-from .satp3 import SATPolicy as TSATPolicy
+from .tsat import TSATPolicy
+from .satp3 import SATP3Policy
 from .satp1 import SATP1Policy
